requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blis==1.1.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
Jinja2==3.1.5
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
numpy==2.2.1
packaging==24.2
pdfminer.six==20240706
preshed==3.0.9
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.41.3
thinc==8.3.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
